Change translations (#3)
The finishings are not supposed to be task names, but the actual finishing itself. For example "hole" really describes the actual hole in the product, not the process of hole making. Thus the correct translation is "Bohrung", not the verb "bohren". That will be used for the work step names.